requirements.txt
asgiref==3.5.0
blinker==1.4
Django==4.0.6
django-filter==21.1
django-rest-framework-mongoengine==3.4.1
djangorestframework==3.13.1
dnspython==2.2.1
Markdown==3.3.6
mongoengine==0.24.1
Pillow==9.1.1
pymongo==4.1.1
python-dateutil==2.8.2
python-snappy==0.6.1
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
psycopg2-binary==2.9.3
django-allauth==0.51.0
django-cors-headers==3.13.0
python-dotenv==0.20.0
dj-rest-auth==2.2.4